# src/DubbingPlatform.Api/OpenApi/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dubbing Platform API",
    "version": "v1",
...
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API root (relative; tenant comes from the JWT tid claim)."
    }
  ],
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "getMe",
        "tags": ["me"],
        "summary": "Current identity with permissions",
        "description": "Returns tenant/user identity plus resolved UX-hint permission strings (never a security boundary; endpoints re-authorize). Includes the caller's IANA `timezone` (default `UTC`) so the frontend can render timestamps without a second preferences round-trip (Plan B 9.1).",
        "responses": {
          "200": {
            "description": "Identity and permissions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/me/preferences": {
      "get": {
        "operationId": "getPreferences",
        "tags": ["me"],
        "summary": "Read identity preferences",
        "description": "Last-writer-wins per key. Unknown keys read as null (never 404).",
        "responses": {
          "200": {
            "description": "Preference map.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "operationId": "updatePreferences",
        "tags": ["me"],
        "summary": "Write identity preferences",
        "description": "Naturally idempotent PUT (no Idempotency-Key). Unknown keys are 400 PREFERENCE_KEY_UNKNOWN; oversize values are 413 PREFERENCE_VALUE_TOO_LARGE.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesResponse"
              },
              "example": {
                "locale": "en",
                "theme": "dark"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored preference map.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "413": {
            "description": "Preference value too large.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "PREFERENCE_VALUE_TOO_LARGE",
                    "message": "Preference value exceeds the size limit.",
                    "correlationId": "corr-01JABCDEF",
                    "details": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT bearer with tid (tenant), sub (user), roles[] claims."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Validation failure (400) with a catalog code.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "VALIDATION_FAILED",
                "message": "Request validation failed.",
                "correlationId": "corr-01JABCDEF",
                "details": {}
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials (401).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication is required.",
                "correlationId": "corr-01JABCDEF",
                "details": {}
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "description": "Frozen public error-code catalog (65 codes). Status mapping: VALIDATION_FAILED 400, UNAUTHORIZED/INVALID_CREDENTIALS/TOKEN_EXPIRED/TOKEN_REUSED 401, FORBIDDEN/CONSENT_REQUIRED/POLICY_DENIED/USER_DISABLED/VOICE_CONSENT_REQUIRED 403, NOT_FOUND/PROJECT_NOT_FOUND/VERSION_NOT_FOUND/VOICE_NOT_FOUND/ARTIFACT_UNAVAILABLE 404, MEDIA_UNSUPPORTED 415, MEDIA_CORRUPT/QC_BLOCKED/ARTIFACT_CHECKSUM_MISMATCH/VOICE_INCOMPATIBLE/IDEMPOTENCY_KEY_REUSED 422, PROVIDER_RATE_LIMITED/PROVIDER_QUOTA_EXHAUSTED/QUOTA_EXCEEDED/RATE_LIMITED 429, RESOURCE_EXHAUSTED/STORAGE_UNAVAILABLE 503, PROVIDER_TIMEOUT 504, PROVIDER_INVALID_RESPONSE/PROVIDER_FAILED 502, PROVIDER_CONFIGURATION_ERROR/PIPELINE_INVARIANT_VIOLATION/INTERNAL_ERROR 500, PREFERENCE_VALUE_TOO_LARGE 413, URL_EXPIRED 410, all other *_CONFLICT/*_ACTIVE/*_LOCKED/SELECTION_CONFLICT/LEASE_LOST/MANUAL_REVIEW_REQUIRED/EXPORT_NOT_READY/EXPORT_INCOMPLETE/OUTPUT_INCOMPLETE/SEGMENT_RETRY_ACTIVE/REVIEW_*_RESOLVED 409, UPLOAD_INCOMPLETE/LANGUAGE_IMMUTABLE/VERSION_SEGMENT_MISMATCH/SEGMENT_TEXT_EMPTY/PREVIEW_TEXT_INVALID/REVIEW_REASON_REQUIRED/REVIEW_EDIT_EMPTY/PREFERENCE_KEY_UNKNOWN/IDEMPOTENCY_KEY_REQUIRED/DUPLICATE_MEDIA 400-or-409 per catalog, TENANT_REQUIRED 401.",
        "enum": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "MEDIA_UNSUPPORTED",
          "MEDIA_CORRUPT",
          "UPLOAD_INCOMPLETE",
          "DUPLICATE_MEDIA",
          "PROVIDER_CONFIGURATION_ERROR",
          "PROVIDER_RATE_LIMITED",
          "PROVIDER_TIMEOUT",
          "PROVIDER_INVALID_RESPONSE",
          "PROVIDER_QUOTA_EXHAUSTED",
          "PROVIDER_FAILED",
          "QC_BLOCKED",
          "QUOTA_EXCEEDED",
          "RATE_LIMITED",
          "RESOURCE_EXHAUSTED",
          "ARTIFACT_UNAVAILABLE",
          "ARTIFACT_CHECKSUM_MISMATCH",
          "LEASE_LOST",
          "PIPELINE_INVARIANT_VIOLATION",
          "MANUAL_REVIEW_REQUIRED",
          "EXPORT_NOT_READY",
          "STORAGE_UNAVAILABLE",
          "CONSENT_REQUIRED",
          "POLICY_DENIED",
          "INTERNAL_ERROR",
          "INVALID_CREDENTIALS",
          "TOKEN_EXPIRED",
          "TOKEN_REUSED",
          "USER_DISABLED",
          "PREFERENCE_KEY_UNKNOWN",
          "PREFERENCE_VALUE_TOO_LARGE",
          "TENANT_REQUIRED",
          "LANGUAGE_IMMUTABLE",
          "SETTINGS_LOCKED_ACTIVE_RUN",
          "PROJECT_NOT_FOUND",
          "PROJECT_HAS_ACTIVE_RUN",
          "SETTINGS_VERSION_CONFLICT",
          "IDEMPOTENCY_KEY_REQUIRED",
          "IDEMPOTENCY_KEY_REUSED",
          "PROJECT_ARCHIVED",
          "RUN_ALREADY_TERMINAL",
          "RUN_ALREADY_ACTIVE",
          "CONFIG_CHANGED_SINCE_RUN",
          "SELECTION_CONFLICT",
          "VERSION_NOT_FOUND",
          "VERSION_SEGMENT_MISMATCH",
          "SEGMENT_TEXT_EMPTY",
          "SEGMENT_RETRY_ACTIVE",
          "VOICE_INCOMPATIBLE",
          "VOICE_CONSENT_REQUIRED",
          "PREVIEW_QUOTA_EXCEEDED",
          "VOICE_NOT_FOUND",
          "PREVIEW_TEXT_INVALID",
          "REVIEW_VERSION_CONFLICT",
          "REVIEW_REASON_REQUIRED",
          "REVIEW_ALREADY_RESOLVED",
          "REVIEW_NOT_RESOLVED",
          "REVIEW_EDIT_EMPTY",
          "EXPORT_INCOMPLETE",
          "OUTPUT_INCOMPLETE",
          "URL_EXPIRED"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Uniform error envelope: { error: { code, message, correlationId, details } }. details carries field errors only; 500s carry a generic message.",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "correlationId", "details"],
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string"
              },
              "correlationId": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "required": ["userId", "tenantId", "permissions"],
        "properties": {
          "userId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "12 UX-hint permission strings (never a security boundary)."
          },
          "timezone": {
            "type": "string",
            "description": "IANA timezone from the caller's `timezone` preference; defaults to `UTC` when unset or unusable (Plan B 9.1)."
          }
        }
      },
      "PreferencesResponse": {
        "type": "object",
        "description": "Identity preference map (last-writer-wins per key).",
        "additionalProperties": {
          "type": "string"
        }
      },
...
    }
  }
}